This was the state of things Russell and I presented at Nashville's Global Game Jam which happened from January's 26 to 28.
On 26 we were presented with a theme: "make me laugh"

From all different ideas presented at the first night, Russell came
with the idea of Mario 64 opening scene where the player can manipulate
the different vertices of iconic plumber's face.  He also proposed to
make the character laugh when touched around the mouth. Since I played
with OpenGL during the week I thought I could contribute to it.

During the first night I iterated over all glfw's examples and found
`boing.c` already had a round shape I could use as a face. I then
decided to do the smiley face out of the sphere already rendered.
I then realized the current shape was not granular enough. I decided to
increase the number of vertices being drawn. I then gave `i` and `j`
integer coordinates to each vertex to be able to distinguish them from
the rest of the face's shape and colored them as black (eyes and months)

On 27's morning I started manipulating some of the vertices' coordinates
to make them interact with the cursor's movement: When the cursor's left
button is pressed and held, new vertices around smiley's mouth and eyes
are dynamically created and their coordinates are calculated based on the
current cursor's Y coordinate in relation to where it was when the button
was initially pressed. When the button is released the new vertices are
destroyed.

Russell worked on generating a sawtooth wave to represent the character's
laugh. We then started working on how to dynamically generate the wave
based on JUCE C++ audio framework. After few attempts we were able to
produce his desired result using the platform.

The exercise also helped me understanding how the framework put things
together and I thought it would be challenging to integrate GLFW's model
into JUCE's given the aggressive time frame we had.

I then tried to write straight to the sound card based on an ALSA's
example I found on-line but that did not work. I then checked
PulseAudio's source code to find the `sync-player.c` test and after few
attempts I was able to produce a sine wave.

I then integrated the code into the smiley face but to my surprise the
PulseAudio's and GLFW's threading model did not perfectly matched and
the audio was all buffered at the beginning making it impossible to
react to the mouse's clicks. I then struggled for few hours on how to
properly integrate audio and graphics, to then finally find an example
of PulseAudio's multithreading interaction on Stack Overflow called
`pcm-player.c`

That example matched perfectly what I was trying to achieve as the
PulseAudio's client is constantly requested for more data rather than
pre-buffering all stream's content. After integrating both codebases I
noticed that the mouse's clicks and PulseAudio's requests for more data
were running in different threads and therefore not in sync.

On 28's morning I found the PulseAudio's sample setting which indicates
the size of the audio buffer and making that shorter considerably improved
the click's reaction time.

I then switched to the wave's generation problem. Thanks to Russell's
expertise on sound engineering he was able to explain me what a sawtooth
wave is, how it is generated, the formula behind it as well as how it is
represented as a linear graph.

I noticed my implementation was not correct and tried few fixes. I then
realized that no matter how I computed the sample's frame there was
always a noise, which indicated the function was not correct.

Once again Russell's discussions about how the wave is supposed to be
helped me to realize it gracefully matches how `x86`'s registers overflow.
The sawtooth wave's format is similar to the modulo's operation, it caps
the ceiling back to to the floor and then it starts over again.

Producing real-time audio is challenging because the listener may caught
any missed frames. I then had to take several factors into consideration
as to how long each frame would take to compute and what happened if there
were not enough resources to keep up with the requirements.

Sound wave generation is very sensitive to time and any glitches may have
an impact into the ending result. I considered calculating the time
based on wall clock for every buffer instead of keeping an incremental
counter for the timer, but that proved to be unnecessary.

The ideal threading model would be to have your frame's producer running
into a separate thread pinned to one specific core and have that code
optimized and limited to constantly consume a certain amount of CPU.
Buffer copies and other CPU contentions should also be avoided.

I then opted to change the sample format from a 4-byte floating point
which has nonprecise arithmetics in their digital representation to a
single byte `uint8_t` data. That considerably reduced the frequency range
I could achieve but I feel it was the right choice in terms of correctness.
Indeed the produced sound is a pure sawtooth wave and it runs without flaw.

The results were presented to the Nashville Game Developers group at the
28's evening.

It was later tested on `aarch64` platform running Arch Linux.

TODO: The code was *hacked* together given the aggressive time frame and
is far from what professional C++ code should look like. I intend to
refactor it in the next couple of weeks adding proper comments.